+++ src/main/scala/nl/vroste/zio/kinesis/client/zionative/fetcher/EnhancedFanOutFetcher.scala
@@ -2,9 +2,10 @@ package nl.vroste.zio.kinesis.client.zionative.fetcher
 
 import nl.vroste.zio.kinesis.client.AdminClient.StreamDescription
 import nl.vroste.zio.kinesis.client.Client.ShardIteratorType
+import nl.vroste.zio.kinesis.client.zionative.Consumer.childShardToShard
 import nl.vroste.zio.kinesis.client.zionative.Fetcher.EndOfShard
 import nl.vroste.zio.kinesis.client.zionative.{ DiagnosticEvent, FetchMode, Fetcher }
-import nl.vroste.zio.kinesis.client.{ childShardToShard, Client, Util }
+import nl.vroste.zio.kinesis.client.{ Client, Util }
 import software.amazon.awssdk.services.kinesis.model.{ ConsumerStatus, ResourceInUseException }
 import zio._
 import zio.clock.Clock
